added two subcomponents to adminDashboard component in order to make it more organized and readable
