+++ src/components/AdminDashboard/DashboardTop.scss
@@ -1,0 +1,43 @@
+#admin-dashboard-title {
+  margin-bottom: 2%;
+}
+
+.statboxes {
+  display: flex;
+  flex-direction: row;
+  justify-content: start;
+}
+
+.statbox {
+  border: 2px solid #e3e8ee;
+  padding: 2%;
+  width: 99%;
+  color: #3c4257;
+  border-radius: 5px;
+  font-weight: 400;
+  margin-right: 20px;
+  margin-bottom: 1%;
+  background-image: linear-gradient(to right, rgb(247, 250, 252), white);
+}
+
+.confirmation-message {
+  width: 90%;
+  text-align: center;
+}
+
+.incidents-wrap {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 2%;
+}
+
+#create-incident-button {
+  border: 1px solid #e3e8ee;
+  background-color: rgb(0, 55, 103);
+  border-width: 2px;
+  border-style: outset;
+  border-radius: 7.5px;
+  color: white;
+  height: 45px;
+}
