--- src/components/AdminDashboard/Edit.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Edit = () => {
-  return <div></div>;
-};
-
-export default Edit;
