+++ src/components/AdminDashboard/Incidents.scss
@@ -1,0 +1,125 @@
+.dashboard-top-flex {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.dashboard-top-flex div {
+  margin: auto 1%;
+}
+
+.dashboard-top-input {
+  display: flex;
+  justify-content: start;
+  align-items: center;
+  width: 7%;
+}
+
+.dashboard-top-input input {
+  margin-left: 2%;
+}
+
+#select-all-label {
+  margin-left: 5%;
+}
+
+.dashboard-top-approve-reject {
+  display: flex;
+  width: 22%;
+}
+
+.dashboard-top-approve-reject button {
+  margin: auto 2%;
+  width: 45%;
+}
+
+.approve-reject-select {
+  border: 1px solid #e3e8ee;
+  background-color: rgb(255, 255, 255);
+  border-width: 2px;
+  border-style: outset;
+  border-radius: 5px;
+  margin-right: 2%;
+  margin-bottom: 1%;
+  color: rgb(60, 66, 87);
+  height: 4%;
+}
+
+.hidden-button {
+  display: none;
+}
+
+.confirmation-message {
+  width: 90%;
+  text-align: center;
+}
+
+.transparent {
+  opacity: 0;
+}
+
+.dashboard-top-page-number {
+  display: flex;
+  width: 71%;
+  justify-content: flex-end;
+}
+
+#items-per-page {
+  margin-left: 2%;
+}
+
+.items-pp-select {
+  margin-left: 2%;
+}
+
+.column-headers {
+  display: flex;
+  justify-content: space-between;
+  border: solid #e3e8ee 1px;
+  width: 100%;
+  align-items: center;
+  margin-bottom: 0px;
+  border-radius: 5px;
+}
+
+.hidden-input {
+  opacity: 0;
+  text-align: left;
+  margin: auto 1%;
+}
+
+.column-headers-flex {
+  display: flex;
+  text-align: left;
+  justify-content: flex-start;
+  width: 100%;
+  align-items: center;
+  color: #3c4257;
+  margin: 0 auto;
+}
+
+.column-headers div h4 {
+  font-weight: bold;
+  padding-top: 0.5%;
+}
+
+.description {
+  margin-left: 0px;
+  text-align: left;
+  width: 45%;
+}
+
+.location {
+  width: 18%;
+  margin-left: 2%;
+}
+
+.date {
+  width: 15%;
+  margin-left: 7%;
+}
+
+.incidents {
+  display: flex;
+  flex-direction: column;
+}
