+++ src/index.scss
@@ -15,7 +15,8 @@
 @import './components/AdminDashboard/CompleteIncident.scss';
 @import './components/AdminDashboard/PendingIncident.scss';
 @import './components/AdminDashboard/AddIncident.scss';
-@import './components/AdminDashboard/Source.scss';
+@import './components/AdminDashboard/DashboardTop.scss';
+@import './components/AdminDashboard/Incidents.scss';
 
 .Map {
   position: relative;
